refactor(pi-agents-tmux): split subagent dispatch wiring
